Modificación final de tipo de datos y ajuste de tamaño de las columnas que guardan URLs

+++ database/migrations/2020_10_27_193142_create_contactos_table.php
@@ -26,7 +26,7 @@ class CreateContactosTable extends Migration
             $table->bigInteger('N_TelefonoFijo')->nullable();
             $table->bigInteger('N_TelefonoMovil')->nullable();
             $table->string('S_Email',255);
-            $table->unsignedInteger('tw_corporativos_id');
+            $table->unsignedBigInteger('tw_corporativos_id');
             $table->timestamps();
             // $table->softDeletes();
